Refactorizando la arquitectura del frontend
