+++ client/src/core/icons/HiddenIcon.vue
@@ -1,5 +1,5 @@
 <template>
-    <svg class="text-complementary-1" width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <svg class="text-complementary-1 select-none" width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
         <path d="M21.4395 0.439499L23.5605 2.5605L19.575 6.546C19.811 6.782 20.0325 7.0155 20.2395 7.2465C21.3213 8.4575 22.2591 9.7899 23.034 11.217L23.0745 11.2935L23.085 11.3175L23.0895 11.325L23.091 11.328L23.427 12L23.091 12.6735L23.0895 12.675L23.085 12.6825L23.0745 12.7065C23.0145 12.8228 22.952 12.9379 22.887 13.0515C22.1411 14.3772 21.2534 15.618 20.2395 16.752C18.4545 18.738 15.6405 21 12 21C9.98995 21 8.23345 20.31 6.76195 19.359L2.56195 23.5605L0.439453 21.4395L21.4395 0.439499ZM14.8995 11.2245L11.223 14.8995C11.7316 15.0346 12.2668 15.0337 12.775 14.8969C13.2831 14.7601 13.7465 14.4923 14.1186 14.1201C14.4907 13.748 14.7586 13.2847 14.8954 12.7765C15.0321 12.2683 15.0345 11.7331 14.8995 11.2245ZM14.703 3.4245C13.8294 3.14373 12.9175 3.00052 12 3C8.35945 3 5.54695 5.262 3.76045 7.2465C2.67857 8.4575 1.74081 9.7899 0.965953 11.217L0.926953 11.2935L0.914953 11.3175L0.910453 11.325C0.562453 12.0195 0.773953 11.5995 0.806953 11.532L0.572953 12L0.908953 12.6705L0.910453 12.675L0.914953 12.6825L0.926953 12.7065C0.986441 12.8228 1.04846 12.9379 1.11295 13.0515C1.58138 13.8843 2.1061 14.6842 2.68345 15.4455L14.703 3.4245Z" fill="#4B4B4B"/>
         </svg>        
 </template>
