+++ client/src/core/icons/ShowIcon.vue
@@ -1,5 +1,5 @@
 <template>
-    <svg class="text-secondary" width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <svg class="text-secondary select-none" width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
         <path d="M14 12C14 12.5304 13.7893 13.0391 13.4142 13.4142C13.0391 13.7893 12.5304 14 12 14C11.4696 14 10.9609 13.7893 10.5858 13.4142C10.2107 13.0391 10 12.5304 10 12C10 11.4696 10.2107 10.9609 10.5858 10.5858C10.9609 10.2107 11.4696 10 12 10C12.5304 10 13.0391 10.2107 13.4142 10.5858C13.7893 10.9609 14 11.4696 14 12Z" fill="currentColor"/>
         <path fill-rule="evenodd" clip-rule="evenodd" d="M12.0001 3C6.40805 3 1.71005 6.824 0.378052 12C1.71005 17.176 6.40805 21 12.0001 21C17.5921 21 22.2901 17.176 23.6221 12C22.2901 6.824 17.5921 3 12.0001 3ZM16.0001 12C16.0001 13.0609 15.5786 14.0783 14.8285 14.8284C14.0783 15.5786 13.0609 16 12.0001 16C10.9392 16 9.92177 15.5786 9.17162 14.8284C8.42148 14.0783 8.00005 13.0609 8.00005 12C8.00005 10.9391 8.42148 9.92172 9.17162 9.17157C9.92177 8.42143 10.9392 8 12.0001 8C13.0609 8 14.0783 8.42143 14.8285 9.17157C15.5786 9.92172 16.0001 10.9391 16.0001 12Z" fill="currentColor"/>
     </svg>
