+++ client/src/router/index.ts
@@ -11,7 +11,7 @@ const router = createRouter({
     {
       path: '/sign-in',
       name: 'sign-in',
-      component: () => import('../module/auth/presentation/views/SignInView.vue')
+      component: () => import('../authentication/views/SignInView.vue')
     }
   ],
 })
